Move `generate_assign_concessions_action` function
Move the `generate_assign_concessions_action` function to the concession helper script because it's no longer going to be exclusive to the revolution phase.
